+++ reptar/calculators/xtb_workers.py
@@ -20,7 +20,12 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
 
+import os
+import subprocess
+from tempfile import NamedTemporaryFile, TemporaryDirectory
 import numpy as np
+from ..writers.xyz import write_xyz
+from ..utils import parse_xyz
 from ..logger import ReptarLogger
 
 log = ReptarLogger(__name__)
@@ -34,7 +39,7 @@ except ImportError:
 
 
 def xtb_python_engrad(
-    idxs, Z, R, charge=0, mult=1, calc_acc=0.1, max_iterations=300, params=None
+    idxs, Z, R, charge=0, mult=1, acc=0.1, max_iterations=300, params=None
 ):
     r"""Ray remote function for computing total electronic energy and atomic
     gradients using xtb.
@@ -53,7 +58,7 @@ def xtb_python_engrad(
         Total molecular charge.
     mult : :obj:`int`, default: ``1``
         Total multiplicity.
-    calc_acc : :obj:`int`, default: ``0.1``
+    acc : :obj:`int`, default: ``0.1``
         Numerical accuracy for calculation. For more information, visit the
         `documentation <https://xtb-python.readthedocs.io/en/latest/
         general-api.html#xtb.interface.Calculator.set_accuracy>`__.
@@ -85,7 +90,7 @@ def xtb_python_engrad(
         params = Param.GFN2xTB
     for i, r in enumerate(R):
         calc = Calculator(params, Z, r, charge, n_upair_ele)
-        calc.set_accuracy(calc_acc)
+        calc.set_accuracy(acc)
         calc.set_max_iterations(max_iterations)
         res = calc.singlepoint()
         g = res.get_gradient()
@@ -93,3 +98,107 @@ def xtb_python_engrad(
         G[i] = g
         E[i] = res.get_energy()
     return idxs, E, G
+
+
+def xtb_opt(
+    idxs,
+    Z,
+    R,
+    input_lines,
+    acc=0.1,
+    n_cores=1,
+    xtb_path="xtb",
+    work_dir=None,
+):
+    r"""Ray remote function for computing total electronic energy and atomic
+    gradients using xtb.
+
+    Parameters
+    ----------
+    idxs : :obj:`numpy.ndarray`, ndim: ``1``
+        Indices of the structures from ``R`` to compute energies and gradients
+        for.
+    Z : :obj:`numpy.ndarray`, ndim: ``1``
+        Atomic numbers of the atoms with respect to ``R``.
+    R : :obj:`numpy.ndarray`, ndim: ``3``
+        Cartesian coordinates of all structures in group. This includes
+        unused structures.
+    input_lines : :obj:`list`
+        Lines for xTB input file. Must at least include ``$chrg`` and ``$spin`` blocks.
+    acc : :obj:`float`, default: ``0.1``
+        Numerical accuracy for calculation. For more information, visit the
+        `documentation <https://xtb-docs.readthedocs.io/en/latest/sp.html#accuracy>`__.
+    n_cores : :obj:`int`, default ``1``
+        Number of cores to use for xTB calculation.
+    xtb_path : :obj:`str`, default: ``"xtb"``
+        Path to xtb executable to use. Defaults to assuming ``xtb`` is in your path.
+    work_dir : :obj:`str`, default: ``None``
+        Work directory for the xtb calculations. If nothing is specified, the
+        temporary directory is used.
+
+    Returns
+    -------
+    :obj:`numpy.ndarray`
+        ``idxs``
+    :obj:`numpy.ndarray`
+        If the optimizations converged or not.
+    :obj:`numpy.ndarray`
+        Optimized geometries.
+    :obj:`numpy.ndarray`
+        Total electronic energies of optimized structures. Units of Hartree.
+    """
+    # pylint: disable=consider-using-with
+    R = R[idxs]
+    opt_conv = np.full(R.shape[0], False, dtype=np.bool8)
+    R_opt = np.empty(R.shape, dtype=np.float64)
+    E = np.zeros(R.shape[0])
+
+    if work_dir is None:
+        temp_dir = TemporaryDirectory()
+        work_dir = temp_dir.name
+    cwd_path = os.getcwd()
+    os.chdir(work_dir)
+
+    xtb_input = NamedTemporaryFile(suffix=".in")
+    with open(xtb_input.name, "w", encoding="utf-8") as f:
+        f.writelines(input_lines)
+
+    for i, r in enumerate(R):
+        # Write temporary input file for coordinates
+        xyz_input = NamedTemporaryFile(suffix=".xyz")
+        write_xyz(xyz_input.name, Z, r)
+
+        output_file = NamedTemporaryFile(suffix=".out")
+        output_path = output_file.name
+
+        xtb_command = [
+            xtb_path,
+            xyz_input.name,
+            "--acc",
+            str(acc),
+            "--input",
+            xtb_input.name,
+            "--opt",
+            "--parallel",
+            str(n_cores),
+        ]
+
+        with open(output_path, "w", encoding="utf-8") as f_out:
+            subprocess.run(xtb_command, check=False, shell=False, stdout=f_out)
+
+        _, comments, r_opt = parse_xyz("xtbopt.xyz")
+        e = float(comments[0].split()[1])
+
+        r_opt_conv = False
+        with open(output_path, "r", encoding="utf-8") as f_out:
+            for line in reversed(list(f_out)):
+                if "*** GEOMETRY OPTIMIZATION CONVERGED AFTER" in line:
+                    r_opt_conv = True
+                    break
+        r_opt = np.array(r_opt)[0]
+
+        opt_conv[i] = r_opt_conv  # pylint: disable=used-before-assignment
+        R_opt[i] = r_opt
+        E[i] = e  # pylint: disable=used-before-assignment
+    os.chdir(cwd_path)
+    return idxs, opt_conv, R_opt, E