+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- `xtb` optimization worker that can be used with ray.
+
 ### Changed
 
 - Drivers and workers should only return guaranteed values.